refactor(schemas): deduplicate apiKeys (canonical at schemas/api-keys/<dialect>.ts)
The Drizzle apiKeys table now lives in schemas/api-keys/{postgres,mysql,
sqlite}.ts with a dialect-aware barrel exposing apiKeyTables(dialect).
database/schema/<dialect>.ts re-exports from the new location so the
apiKeysRelations() block keeps compiling.

Plan A originally framed this task as deduplicating a Drizzle table that
appeared at both database/schema/<dialect>.ts and schemas/api-keys.ts. The
top-level schemas/api-keys.ts turned out to contain Zod validators only
(ApiKeyTokenTypeSchema, CreateApiKeySchema, UpdateApiKeySchema, etc.), not
a duplicate Drizzle table. To match the Task 4 convention for Zod files,
the Zod module was renamed to schemas/_zod/api-keys.ts and the single
direct importer (packages/nextly/src/api/api-keys.ts) was updated. The
_zod/ barrel re-exports it so `import { CreateApiKeySchema } from "@nextly/schemas"`
keeps working.
