+++ packages/nextly/src/api/api-keys.ts
@@ -29,7 +29,10 @@ import { container } from "../di";
 import type { ApiKeyService } from "../domains/auth/services/api-key-service";
 import { NextlyError } from "../errors/nextly-error";
 import { getCachedNextly } from "../init";
-import { CreateApiKeySchema, UpdateApiKeySchema } from "../schemas/api-keys";
+import {
+  CreateApiKeySchema,
+  UpdateApiKeySchema,
+} from "../schemas/_zod/api-keys";
 import { isSuperAdmin } from "../services/lib/permissions";
 
 import { readJsonBody } from "./read-json-body";
