+++ packages/nextly/src/database/schema/sqlite.ts
@@ -109,51 +109,11 @@ import {
  * - Safe expression evaluation (sandboxed)
  */
 
-/**
- * API Keys table for programmatic API authentication (SQLite).
- *
- * See postgres.ts for full documentation. Main differences:
- * - Uses TEXT for all string columns (SQLite has no varchar length enforcement)
- * - Uses INTEGER { mode: "timestamp" } for all datetime columns
- * - Uses INTEGER { mode: "boolean" } for boolean columns
- * - JSON stored as TEXT where applicable
- */
-export const apiKeys = sqliteTable(
-  "api_keys",
-  {
-    id: text("id").primaryKey(),
-    name: text("name").notNull(),
-    description: text("description"),
-    // SHA-256 hex digest — primary lookup column, never the raw key
-    keyHash: text("key_hash").notNull(),
-    // First 16 characters of the full key for display
-    keyPrefix: text("key_prefix").notNull(),
-    tokenType: text("token_type").notNull(),
-    // onDelete: "set null" — deleted role makes key permission-less (safe 403)
-    roleId: text("role_id").references(() => roles.id, {
-      onDelete: "set null",
-    }),
-    // onDelete: "cascade" — deleting a user removes all their keys
-    userId: text("user_id")
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    expiresAt: integer("expires_at", { mode: "timestamp" }),
-    lastUsedAt: integer("last_used_at", { mode: "timestamp" }),
-    isActive: integer("is_active", { mode: "boolean" }).notNull().default(true),
-    createdAt: integer("created_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-    updatedAt: integer("updated_at", { mode: "timestamp" })
-      .notNull()
-      .$defaultFn(() => new Date()),
-  },
-  t => [
-    uniqueIndex("api_keys_key_hash_unique").on(t.keyHash),
-    index("api_keys_user_id_idx").on(t.userId),
-    index("api_keys_role_id_idx").on(t.roleId),
-    index("api_keys_is_active_expires_at_idx").on(t.isActive, t.expiresAt),
-  ]
-);
+// apiKeys — moved to schemas/api-keys/sqlite.ts (Plan A Task 11).
+// Re-exported here so existing consumers and relations() blocks below keep working
+// until Task 16 sweeps imports to @nextly/schemas.
+export { apiKeys } from "../../schemas/api-keys/sqlite";
+import { apiKeys } from "../../schemas/api-keys/sqlite";
 
 // activityLog moved to schemas/audit/sqlite.ts (Plan A Task 9) and re-exported
 // with auditLog above.
