+++ packages/nextly/src/schemas/_zod/index.ts
@@ -11,3 +11,4 @@
 export * from "./user";
 export * from "./rbac";
 export * from "./validation";
+export * from "./api-keys";
